--- a/transport-native-io_uring/src/main/java/io/netty/channel/uring/IOUringSubmissionQueue.java
+++ b/transport-native-io_uring/src/main/java/io/netty/channel/uring/IOUringSubmissionQueue.java
@@ -136,9 +136,8 @@ final class IOUringSubmissionQueue {
             }
         }
 
-        PlatformDependent.putByte(sqe + SQE_FLAGS_FIELD, (byte) 0);
-
-
+        // TODO: Make it configurable if we should use this flag or not.
+        PlatformDependent.putByte(sqe + SQE_FLAGS_FIELD, (byte) Native.IOSQE_ASYNC);
 
         // pad field array -> all fields should be zero
         long offsetIndex = 0;