@normanmaurer 2020-09-04
/cc @njhill @franz1981 @chrisvest ... 

Special thanks to @chrisvest for the suggestion.

NIO as baseline:

```
norman@ubuntu-dev:~/Apps/rust_echo_bench$ ./run.sh 9090
Linux ubuntu-dev 5.8.5-iouringfix+ #2 SMP Thu Sep 3 12:58:01 CEST 2020 x86_64 x86_64 x86_64 GNU/Linux
pid 126331's current affinity list: 0
pid 126331's new affinity list: 0
    Finished release [optimized] target(s) in 0.01s
     Running `target/release/echo_bench --address 'localhost:9090' --number 500 --duration 30 --length 128`
Benchmarking: localhost:9090
500 clients, running 128 bytes, 30 sec.
Speed: 78744 request/sec, 78744 response/sec
Requests: 2362321
Responses: 2362320
```

Before this PR:

```
norman@ubuntu-dev:~/Apps/rust_echo_bench$ ./run.sh 8080
Linux ubuntu-dev 5.8.5-iouringfix+ #2 SMP Thu Sep 3 12:58:01 CEST 2020 x86_64 x86_64 x86_64 GNU/Linux
pid 138209's current affinity list: 0
pid 138209's new affinity list: 0
   Finished release [optimized] target(s) in 0.00s
    Running `target/release/echo_bench --address 'localhost:8080' --number 500 --duration 30 --length 128`
Benchmarking: localhost:8080
500 clients, running 128 bytes, 30 sec.
Speed: 124256 request/sec, 124256 response/sec
Requests: 3727709
Responses: 3727708
```

With this change:

```
norman@ubuntu-dev:~/Apps/rust_echo_bench$ ./run.sh 8080
Linux ubuntu-dev 5.8.5-iouringfix+ #2 SMP Thu Sep 3 12:58:01 CEST 2020 x86_64 x86_64 x86_64 GNU/Linux
pid 143850's current affinity list: 0
pid 143850's new affinity list: 0
    Finished release [optimized] target(s) in 0.00s
     Running `target/release/echo_bench --address 'localhost:8080' --number 500 --duration 30 --length 128`
Benchmarking: localhost:8080
500 clients, running 128 bytes, 30 sec.
Speed: 244279 request/sec, 244279 response/sec
Requests: 7328385
Responses: 7328377
```

🔥