@1Jo1 2020-09-04
my test:

Nio:

cargo run --release -- --address "127.0.0.1:8080" --number 500 --duration 30 --length 128
    Finished release [optimized] target(s) in 0.00s
     Running `target/release/echo_bench --address '127.0.0.1:8080' --number 500 --duration 30 --length 128`
Benchmarking: 127.0.0.1:8080
500 clients, running 128 bytes, 30 sec.

Speed: 139344 request/sec, 139344 response/sec
Requests: 4180346
Responses: 4180346


Use Flag:

cargo run --release -- --address "127.0.0.1:8080" --number 500 --duration 30 --length 128 
    Finished release [optimized] target(s) in 0.01s
     Running `target/release/echo_bench --address '127.0.0.1:8080' --number 500 --duration 30 --length 128`
Benchmarking: 127.0.0.1:8080
500 clients, running 128 bytes, 30 sec.

Speed: 323870 request/sec, 323870 response/sec
Requests: 9716123
Responses: 9716114