@njhill 2020-09-04
Nice one @chrisvest @normanmaurer!

This is interesting and slightly surprising to me since I assumed most of the submissions would be non-blocking. I wonder if setting this selectively based on the operation could help even more (i.e. just for the polls)?