@normanmaurer 2020-09-05
Unfortunally it seems like this one can cause some problems with a few kernels:

https://lore.kernel.org/io-uring/6428c1ee0234105d18c5e3e88aa00c57@nickhill.org/T/#t

But a fix is already proposed.