@njhill 2020-09-05
@normanmaurer fyi simple workaround for now is to change it to:
```java
        PlatformDependent.putByte(sqe + SQE_FLAGS_FIELD,
            (byte) (op != Native.IORING_OP_WRITEV ? Native.IOSQE_ASYNC : 0));
```