@normanmaurer 2020-09-05
@njhill yep thanks... I am about to compile a new kernel with the proposed fix and see if this fixes it :)