@normanmaurer 2020-09-05
For the record here is the proposed fix:

https://lore.kernel.org/io-uring/d8404079-fe7e-3f42-4460-22328b12b0fa@kernel.dk/