@normanmaurer 2020-09-05
@njhill at the end I decided to just do a PR with your workaround for now to make it easier for people to develop until a fix is merged into the kernel: https://github.com/1Jo1/netty/pull/36