@1Jo1 2020-09-05
> This is interesting and slightly surprising to me since I assumed most of the submissions would be non-blocking. I wonder if setting this selectively based on the operation could help even more (i.e. just for the polls)?

@njhill yeah I'm quite surprised too, I guess writeev/write and read operation are faster, it seems to be that IOSQE_ASYNC skips the arm handler and the operations will be executed in async threads from the start, I will test that later :)